Add recovery key unwrap and API integration
Introduces recovery API client (fetchRecoveryParams) and recoveryUnwrap utility for unwrapping vault keys using passphrase-derived KEK. Updates kdfService.client.ts to simplify return type. These changes enable the recovery flow for vault key decryption.

+++ frontend/src/crypto/kdfService.client.ts
@@ -39,5 +39,5 @@ export async function deriveKek256(passphrase: string, params: Argon2idParams): 
   });
 
   // result.hash is Uint8Array
-  return result.hash as Uint8Array;
+  return result.hash;
 }
